requirements.txt
pyzmq==25.1.2
opencv-python==4.9.0.80
numpy==1.26.4